+++ plugins/postcss-logical/.rollup.js
@@ -1,15 +1,15 @@
 import babel from 'rollup-plugin-babel';
 
 export default {
-	input: 'index.js',
+	input: 'src/index.js',
 	output: [
-		{ file: 'index.cjs.js', format: 'cjs', sourcemap: true },
-		{ file: 'index.es.mjs', format: 'es', sourcemap: true }
+		{ file: 'index.cjs.js', format: 'cjs', sourcemap: true, strict: false },
+		{ file: 'index.esm.mjs', format: 'esm', sourcemap: true, strict: false }
 	],
 	plugins: [
 		babel({
 			presets: [
-				['@babel/env', { modules: false, targets: { node: 6 } }]
+				['@babel/env', { modules: false, targets: { node: 8 } }]
 			]
 		})
 	]