+++ plugins/postcss-logical/lib/match-supported-properties.js
@@ -1,1 +1,0 @@
-export default /^(?:(inset|margin|padding)(?:-(block|block-start|block-end|inline|inline-start|inline-end|start|end))|(min-|max-)?(block|inline)-(size))$/i;