+++ plugins/postcss-logical/lib/transform-size.js
@@ -1,8 +1,0 @@
-import matchSize from './match-size';
-
-export default decl => {
-	decl.prop = decl.prop.replace(
-		matchSize,
-		($0, minmax, flow) => `${minmax||''}${'block' === flow ? 'height' : 'width'}`
-	);
-};