+++ plugins/postcss-logical/src/lib/transform-size.js
@@ -1,0 +1,10 @@
+export default {
+	// block-size
+	'block-size': decl => {
+		decl.prop = 'width';
+	},
+	// inline-size
+	'inline-size': decl => {
+		decl.prop = 'height';
+	}
+}